testing/web-platform/mozilla/tests/css/css-forms/input-type-range-thumb-hit-test.html

<!doctype html>
<meta charset="utf-8">
<title>Range thumb should be hit-testable separate from the track even if themed</title>
<link rel=help href="https://bugzilla.mozilla.org/show_bug.cgi?id=1973748">
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<style>
  input[type="range"] {
    pointer-events: none;

    &::-moz-range-thumb {
      pointer-events: auto;
    }
  }
</style>
<input type="range" min="0" max="100" value="50">
<script>
  // The thumb should be in the middle of the input.
  test(function () {
    let input = document.querySelector("input");
    let r = input.getBoundingClientRect();
    assert_equals(document.elementFromPoint(r.left + r.width / 2, r.top + r.height / 2), input, "Should be able to hit the <input>");
  });
</script>
